Moving custom app out of examples
