--- examples/src/index.tsx
@@ -1,20 +1,0 @@
-import * as React from "react";
-import { AppUiProvider } from "@canva/app-ui-kit";
-import { createRoot } from "react-dom/client";
-import { App } from "./app";
-import "@canva/app-ui-kit/styles.css";
-
-const root = createRoot(document.getElementById("root")!);
-function render() {
-  root.render(
-    <AppUiProvider>
-      <App />
-    </AppUiProvider>
-  );
-}
-
-render();
-
-if (module.hot) {
-  module.hot.accept("./app", render);
-}
