+++ src/adapter.ts
